refactor: delete flow node instances tree v1
